requirements.txt
matplotlib==3.2.1
python-telegram-bot==12.7
antlr4-python3-runtime==4.8
pickle-mixin==1.0.2
